requirements.txt
discord.py
PyYaml
psutil
databases
aiohttp

fastapi
starlette
uvicorn
itsdangerous
boto3